requirements.txt
langchain
langchain_core
langchain_community
langchain_google_genai
black
prettier
pydantic
langchain_groq
python-dotenv
aiohttp
rich
